# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Veni AI Sustainability Cockpit API",
    "version": "v1"
  },
  "paths": {
...
    "/catalog/tenants": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Tenants",
        "operationId": "list_tenants_catalog_tenants_get",
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slug"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Create Tenant",
        "operationId": "create_tenant_catalog_tenants_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/tenants/{tenant_id}/projects": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_catalog_tenants__tenant_id__projects_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/projects": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Create Project",
        "operationId": "create_project_catalog_projects_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/bootstrap-workspace": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Bootstrap Workspace",
        "operationId": "bootstrap_workspace_catalog_bootstrap_workspace_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceBootstrapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceBootstrapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/workspace-context": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Workspace Context",
        "operationId": "get_workspace_context_catalog_workspace_context_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Tenant Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrandKitResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name"
          },
          "primary_color": {
            "type": "string",
            "title": "Primary Color"
          },
          "secondary_color": {
            "type": "string",
            "title": "Secondary Color"
          },
          "accent_color": {
            "type": "string",
            "title": "Accent Color"
          },
          "font_family_headings": {
            "type": "string",
            "title": "Font Family Headings"
          },
          "font_family_body": {
            "type": "string",
            "title": "Font Family Body"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "brand_name",
          "primary_color",
          "secondary_color",
          "accent_color",
          "font_family_headings",
          "font_family_body"
        ],
        "title": "BrandKitResponse"
      },
      "CompanyProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "legal_name": {
            "type": "string",
            "title": "Legal Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "headquarters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headquarters"
          },
          "ceo_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ceo Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "legal_name"
        ],
        "title": "CompanyProfileResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreateRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "minLength": 1,
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Code"
          },
          "reporting_currency": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Reporting Currency",
            "default": "TRY"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "name",
          "code"
        ],
        "title": "ProjectCreateRequest"
      },
...
      "ProjectListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ProjectListResponse"
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "reporting_currency": {
            "type": "string",
            "title": "Reporting Currency"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "code",
          "reporting_currency",
          "status"
        ],
        "title": "ProjectResponse"
      },
...
      "TenantCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "TenantCreateRequest"
      },
      "TenantListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TenantResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TenantListResponse"
      },
      "TenantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "status"
        ],
        "title": "TenantResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceBootstrapRequest": {
        "properties": {
          "tenant_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Tenant Name"
          },
          "tenant_slug": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Tenant Slug"
          },
          "project_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Project Name"
          },
          "project_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Project Code"
          },
          "reporting_currency": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Reporting Currency",
            "default": "TRY"
          }
        },
        "type": "object",
        "required": [
          "tenant_name",
          "tenant_slug",
          "project_name",
          "project_code"
        ],
        "title": "WorkspaceBootstrapRequest"
      },
      "WorkspaceBootstrapResponse": {
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantResponse"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectResponse"
          },
          "company_profile": {
            "$ref": "#/components/schemas/CompanyProfileResponse"
          },
          "brand_kit": {
            "$ref": "#/components/schemas/BrandKitResponse"
          },
          "integrations": {
            "items": {
              "$ref": "#/components/schemas/IntegrationConfigSummaryResponse"
            },
            "type": "array",
            "title": "Integrations"
          },
          "blueprint_version": {
            "type": "string",
            "title": "Blueprint Version"
          },
          "tenant_created": {
            "type": "boolean",
            "title": "Tenant Created"
          },
          "project_created": {
            "type": "boolean",
            "title": "Project Created"
          }
        },
        "type": "object",
        "required": [
          "tenant",
          "project",
          "company_profile",
          "brand_kit",
          "integrations",
          "blueprint_version",
          "tenant_created",
          "project_created"
        ],
        "title": "WorkspaceBootstrapResponse"
      },
      "WorkspaceContextResponse": {
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantResponse"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectResponse"
          },
          "company_profile": {
            "$ref": "#/components/schemas/CompanyProfileResponse"
          },
          "brand_kit": {
            "$ref": "#/components/schemas/BrandKitResponse"
          },
          "integrations": {
            "items": {
              "$ref": "#/components/schemas/IntegrationConfigSummaryResponse"
            },
            "type": "array",
            "title": "Integrations"
          },
          "blueprint_version": {
            "type": "string",
            "title": "Blueprint Version"
          }
        },
        "type": "object",
        "required": [
          "tenant",
          "project",
          "company_profile",
          "brand_kit",
          "integrations",
          "blueprint_version"
        ],
        "title": "WorkspaceContextResponse"
      }
    }
  }
}